fix return value from incorrect mask supplied
