+++ Cargo.toml
@@ -40,3 +40,9 @@ grin_wallet_util = { path = "./util", version = "2.1.0-beta.1" }
 
 [build-dependencies]
 built = "0.3"
+
+[dev-dependencies]
+url = "1.7.2"
+serde = "1"
+serde_derive = "1"
+serde_json = "1"
