+++ tests/data/retrieve_info.req.json
@@ -1,0 +1,10 @@
+{
+	"jsonrpc": "2.0",
+	"method": "retrieve_summary_info",
+	"params": {
+		"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+		"refresh_from_node": true,
+		"minimum_confirmations": 1
+	},
+	"id": 1
+}
